% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimacionBrownC.R
\name{estimacionBrownC}
\alias{estimacionBrownC}
\title{Quadratic Brown Estimation}
\usage{
estimacionBrownC(datos, posicion_estimada, alfa)
}
\arguments{
\item{datos}{Array of data with columns t, Zt}

\item{posicion_estimada}{estimation period}

\item{alfa}{parameter value <1}
}
\value{
returns the value of the smoothed Zt of the requested period
}
\description{
Quadratic Brown Estimation
}
