% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suavi_brownC.R
\name{suavi_brownC}
\alias{suavi_brownC}
\title{Brown-Quadratic Smoothing}
\usage{
suavi_brownC(datos, alfa)
}
\arguments{
\item{datos}{Array of data with columns t, Zt}

\item{alfa}{parameter value <1, it can be a vector to test the best}
}
\value{
data table of the smoothed Zt and bt's smoothed and the at
}
\description{
Brown-Quadratic Smoothing
}
