% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuavizaHOLTW.R
\name{SuavizaHOLTW}
\alias{SuavizaHOLTW}
\title{Smoothing Holt-Winter}
\usage{
SuavizaHOLTW(
  datos,
  col_zt,
  anio,
  mes,
  A,
  C,
  D,
  metodo = c("aditivo", "multiplicativo"),
  prediccion = NULL
)
}
\arguments{
\item{datos}{Array of data with columns t, Zt, month and year}

\item{col_zt}{numeric value of the Zt column}

\item{anio}{numeric value of the years column}

\item{mes}{numeric value of the month column}

\item{A}{parameter value <1, A + C + D = 1}

\item{C}{parameter value <1, A + C + D = 1}

\item{D}{value of parameter <1, A + C + D = 1}

\item{metodo}{can take the values "additive" or "multiplicative"}

\item{prediccion}{numerical value of the time (t) or period of which the smoothed Zt is to be predicted}
}
\value{
data table Ft, Zt, Tt, Zt smoothed and the at
}
\description{
Additive or multiplicative smoothing method, which gives the table with the values
of Ft, Zt, Tt, Zt smoothed and the at (error). In addition to obtaining the general deviation
of the smoothed string and predicted values if necessary.
}
